+++ internal/provider/data_source_topic.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	webclient "axual-webclient"
+	"axual.com/terraform-provider-axual/internal/provider/utils"
 	"context"
 	"fmt"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
@@ -32,7 +33,9 @@ type topicDataSourceData struct {
 	Name            types.String `tfsdk:"name"`
 	Description     types.String `tfsdk:"description"`
 	KeyType         types.String `tfsdk:"key_type"`
+	KeySchema       types.String `tfsdk:"key_schema"`
 	ValueType       types.String `tfsdk:"value_type"`
+	ValueSchema     types.String `tfsdk:"value_schema"`
 	Owners          types.String `tfsdk:"owners"`
 	RetentionPolicy types.String `tfsdk:"retention_policy"`
 	Id              types.String `tfsdk:"id"`
@@ -67,6 +70,14 @@ func (d *topicDataSource) Schema(ctx context.Context, req datasource.SchemaReque
 			},
 			"value_type": schema.StringAttribute{
 				MarkdownDescription: "The value type and reference to the schema (if applicable). Read more: https://docs.axual.io/axual/2024.2/self-service/topic-management.html#value-type",
+				Computed:            true,
+			},
+			"key_schema": schema.StringAttribute{
+				MarkdownDescription: "The key schema UID if `key_type` is 'AVRO'.",
+				Computed:            true,
+			},
+			"value_schema": schema.StringAttribute{
+				MarkdownDescription: "The value schema UID if `value_type` is 'AVRO'.",
 				Computed:            true,
 			},
 			"owners": schema.StringAttribute{
@@ -155,5 +166,17 @@ func mapTopicDataSourceResponseToData(ctx context.Context, data *topicDataSource
 		data.Description = types.StringNull()
 	} else {
 		data.Description = types.StringValue(topic.Description.(string))
+	}
+
+	if data.KeyType.ValueString() == "AVRO" {
+		data.KeySchema = utils.SetStringValue(topic.Embedded.KeySchema.Uid)
+	} else {
+		data.KeySchema = types.StringNull()
+	}
+
+	if data.ValueType.ValueString() == "AVRO" {
+		data.ValueSchema = utils.SetStringValue(topic.Embedded.ValueSchema.Uid)
+	} else {
+		data.ValueSchema = types.StringNull()
 	}
 }